requirements.txt
google-generativeai==0.5.1
gspread==6.1.0
python-dotenv~=1.0.1
oauth2client==4.1.3
requests~=2.32.3